rename Gloubal to Global, update background-image, add notFound Component
