+++ client/src/components/Global/Privacy.jsx
