--- client/src/components/Gloubal/NotFound.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function NotFound() {
-  return (
-    <div>NotFound</div>
-  )
-}
-
-export default NotFound
